+++ src/QCArrayPrivate.h
@@ -73,7 +73,7 @@ struct QCArray {
     };
 };
 
-extern void *_QCMallocData(QCArrayDataType type, int count, size_t  *size);
+extern void *_QCMallocData(QCArrayDataType type, size_t count, size_t  *size);
 
 extern void QCArrayXORAt(QCArrayRef array, int index, int value);
